refactor: Move PhoneNumberUIState to its own file
